mtools 3.6

+++ force_io.c
@@ -4,7 +4,7 @@
  * written by:
  *
  * Alain L. Knaff			
- * Alain.Knaff@inrialpes.fr
+ * Alain.Knaff@poboxes.com
  *
  */
 
@@ -13,8 +13,8 @@
 #include "stream.h"
 
 static int force_io(Stream_t *Stream,
-		    char *buf, int start, int len,
-		    int (*io)(Stream_t *, char *, int, int))
+		    char *buf, off_t start, size_t len,
+		    int (*io)(Stream_t *, char *, off_t, size_t))
 {
 	int ret;
 	int done=0;
@@ -35,13 +35,13 @@ static int force_io(Stream_t *Stream,
 	return done;
 }
 
-int force_write(Stream_t *Stream, char *buf, int start, int len)
+int force_write(Stream_t *Stream, char *buf, off_t start, size_t len)
 {
 	return force_io(Stream, buf, start, len,
 			Stream->Class->write);
 }
 
-int force_read(Stream_t *Stream, char *buf, int start, int len)
+int force_read(Stream_t *Stream, char *buf, off_t start, size_t len)
 {
 	return force_io(Stream, buf, start, len,
 			Stream->Class->read);
